Adding US speaker, power and jumper jerky
For the speaker I'm not sure how to represent the different manufacturers, since it's not the same depending on location. 
For the jumper jerky, the manufacturer is unknown. The headers are referring to two types of information -- who made it, and who is providing it. Perhaps this could be fixed with **Manufacturer/Publisher** becoming **Designer** and **Link** becoming **Where to get it**.